pyproject.toml
# GPXMapper Project Configuration
# ------------------------------
# Build:
#   - Development: pip install -e .
#   - Executable: python build_exe.py
#   - Package: python -m build

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gpxmapper"
version = "0.1.0"
description = "GPX to video mapper - creates videos from GPX tracks"
authors = [
    { name = "Developer", email = "dev@example.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gpxpy>=1.5.0",
    "typer>=0.9.0",
    "opencv-python>=4.8.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.scripts]
gpxmapper = "gpxmapper.cli:app"




